Initial work on composable Dot rendering flow.
