+++ scabbard-processor/src/main/java/dev/arunkumar/scabbard/plugin/output/OutputManager.kt
@@ -51,8 +51,7 @@ interface OutputManager {
 /**
  * OutputManager implementation backed by a [Filer]
  */
-interface FilerOutputManager :
-  OutputManager {
+interface FilerOutputManager : OutputManager {
   val filer: Filer
 }
