--- scabbard-processor/src/main/java/dev/arunkumar/scabbard/plugin/parser/BindingExt.kt
@@ -1,8 +1,0 @@
-package dev.arunkumar.scabbard.plugin.parser
-
-import dagger.model.Binding
-
-internal fun Binding.scopeName() = when {
-  scope().isPresent -> scope().get().name
-  else -> null
-}
