--- scabbard-processor/src/main/java/dev/arunkumar/scabbard/plugin/parser/ScopeExt.kt
@@ -1,5 +1,0 @@
-package dev.arunkumar.scabbard.plugin.parser
-
-import dagger.model.Scope
-
-internal inline val Scope.name get() = "@${scopeAnnotationElement().simpleName}"
