+++ scabbard-processor/src/main/java/dev/arunkumar/scabbard/plugin/processor/graphviz/renderer/EntryPointsRenderer.kt
@@ -1,0 +1,30 @@
+package dev.arunkumar.scabbard.plugin.processor.graphviz.renderer
+
+import dagger.model.Binding
+import dagger.model.BindingKind
+import dev.arunkumar.dot.dsl.DotGraphBuilder
+import dev.arunkumar.scabbard.plugin.processor.graphviz.RenderingContext
+
+class EntryPointsRenderer(
+  override val renderingContext: RenderingContext
+) : Renderer<List<Binding>> {
+  override fun DotGraphBuilder.build(renderingElement: List<Binding>) {
+    cluster("Entry Points") {
+      graphAttributes {
+        "labeljust" eq "l"
+        "label" eq "Entry Points"
+      }
+      renderingElement.forEach { node ->
+        val label = when (node.kind()) {
+          BindingKind.MEMBERS_INJECTION -> "inject (${node.label})"
+          else -> node.label
+        }
+        node.id {
+          "shape" eq "component"
+          "label" eq label
+          "penwidth" eq 2
+        }
+      }
+    }
+  }
+}
